LPS-155444 Move functions to common directory
