+++ modules/dxp/apps/osb/osb-site-initializer/osb-site-initializer-customer-portal/extra/remote-app/src/routes/customer-portal/components/CardSubscription/index.js
@@ -14,10 +14,10 @@ import {useState} from 'react';
 import i18n from '../../../../common/I18n';
 import {StatusTag} from '../../../../common/components';
 import {useAppPropertiesContext} from '../../../../common/contexts/AppPropertiesContext';
+import getDateCustomFormat from '../../../../common/utils/getDateCustomFormat';
+import getKebabCase from '../../../../common/utils/getKebabCase';
 import ModalCardSubscription from '../../containers/ModalCardSubscription';
 import {STATUS_TAG_TYPES} from '../../utils/constants';
-import getDateCustomFormat from '../../utils/getDateCustomFormat';
-import getKebabCase from '../../utils/getKebabCase';
 
 const dateFormat = {
 	day: '2-digit',
