+++ modules/dxp/apps/osb/osb-site-initializer/osb-site-initializer-customer-portal/extra/remote-app/src/routes/customer-portal/components/ProjectSupport/index.js
@@ -10,8 +10,8 @@
  */
 
 import i18n from '../../../../common/I18n';
+import getKebabCase from '../../../../common/utils/getKebabCase';
 import {useCustomerPortal} from '../../context';
-import getKebabCase from '../../utils/getKebabCase';
 import SlaCard from '../SlaCard';
 import ProjectSupportSkeleton from './Skeleton';
