+++ modules/dxp/apps/osb/osb-site-initializer/osb-site-initializer-customer-portal/extra/remote-app/src/routes/customer-portal/components/SlaCard/index.js
@@ -15,7 +15,7 @@ import {useEffect, useState} from 'react';
 import i18n from '../../../../common/I18n';
 import {FORMAT_DATE} from '../../../../common/utils/constants/slaCardDate';
 import {SLA_CARD_NAMES} from '../../../../common/utils/constants/slaCardNames';
-import getDateCustomFormat from '../../utils/getDateCustomFormat';
+import getDateCustomFormat from '../../../../common/utils/getDateCustomFormat';
 import SlaCardLayout from './Layout';
 
 const SlaCard = ({project}) => {
@@ -72,18 +72,15 @@ const SlaCard = ({project}) => {
 			slaRawData.current.dateStart = slaRawData.expired.dateStart;
 			slaRawData.current.dateEnd = slaRawData.future.dateEnd;
 			slaFiltedData.push(slaRawData.current);
-		}
-		else if (slaRawData.current.title === slaRawData.expired.title) {
+		} else if (slaRawData.current.title === slaRawData.expired.title) {
 			slaRawData.current.dateStart = slaRawData.expired.dateStart;
 			slaFiltedData.push(slaRawData.current);
 			slaFiltedData.push(slaRawData.future);
-		}
-		else if (slaRawData.current.title === slaRawData.future.title) {
+		} else if (slaRawData.current.title === slaRawData.future.title) {
 			slaRawData.current.dateEnd = slaRawData.future.dateEnd;
 			slaFiltedData.push(slaRawData.current);
 			slaFiltedData.push(slaRawData.expired);
-		}
-		else {
+		} else {
 			slaFiltedData.push(slaRawData.current);
 			slaFiltedData.push(slaRawData.expired);
 			slaFiltedData.push(slaRawData.future);
@@ -104,8 +101,7 @@ const SlaCard = ({project}) => {
 		if (slaData[nextPosition]) {
 			setSlaSelected(slaData[nextPosition].label);
 			setSlaPosition(nextPosition);
-		}
-		else {
+		} else {
 			setSlaSelected(slaData[0].label);
 			setSlaPosition(0);
 		}
