+++ modules/dxp/apps/osb/osb-site-initializer/osb-site-initializer-customer-portal/extra/remote-app/src/routes/customer-portal/components/SubscriptionsFilterByStatus/index.js
@@ -14,8 +14,8 @@ import {DropDown} from '@clayui/core';
 import ClayIcon from '@clayui/icon';
 import React, {useState} from 'react';
 import i18n from '../../../../common/I18n';
+import getKebabCase from '../../../../common/utils/getKebabCase';
 import {SUBSCRIPTIONS_STATUS} from '../../utils/constants';
-import getKebabCase from '../../utils/getKebabCase';
 
 const SubscriptionsFilterByStatus = ({selectedStatus, setSelectedStatus}) => {
 	const [active, setActive] = useState(false);
