+++ modules/dxp/apps/osb/osb-site-initializer/osb-site-initializer-customer-portal/extra/remote-app/src/routes/customer-portal/containers/ActivationKeysTable/utils/getProductName.js
@@ -10,8 +10,8 @@
  */
 
 import i18n from '../../../../../common/I18n';
+import getKebabCase from '../../../../../common/utils/getKebabCase';
 
-import getKebabCase from '../../../utils/getKebabCase';
 import {getPascalCase} from '../../../utils/getPascalCase';
 
 export function getProductName(activationKey) {
