+++ modules/dxp/apps/osb/osb-site-initializer/osb-site-initializer-customer-portal/extra/remote-app/src/routes/customer-portal/containers/GenerateNewKey/GenerateCardLayout/index.js
@@ -11,7 +11,7 @@
 
 import ClayCard from '@clayui/card';
 import i18n from '../../../../../common/I18n';
-import getDateCustomFormat from '../../../utils/getDateCustomFormat';
+import getDateCustomFormat from '../../../../../common/utils/getDateCustomFormat';
 
 const dateFormat = {
 	day: '2-digit',
