+++ modules/dxp/apps/osb/osb-site-initializer/osb-site-initializer-customer-portal/extra/remote-app/src/routes/customer-portal/containers/ModalCardSubscription/index.js
@@ -15,9 +15,9 @@ import React, {useState} from 'react';
 import i18n from '../../../../common/I18n';
 import {Button, StatusTag, Table} from '../../../../common/components';
 import {getAccountSubscriptionsTerms} from '../../../../common/services/liferay/graphql/queries';
+import getDateCustomFormat from '../../../../common/utils/getDateCustomFormat';
+import getKebabCase from '../../../../common/utils/getKebabCase';
 import {STATUS_TAG_TYPES} from '../../utils/constants';
-import getDateCustomFormat from '../../utils/getDateCustomFormat';
-import getKebabCase from '../../utils/getKebabCase';
 
 const dateFormat = {
 	day: '2-digit',
