+++ modules/dxp/apps/osb/osb-site-initializer/osb-site-initializer-customer-portal/extra/remote-app/src/routes/customer-portal/pages/Project/routes/Outlets/ActivationOutlet.js
@@ -19,8 +19,8 @@ import {
 	useResolvedPath,
 } from 'react-router-dom';
 import i18n from '../../../../../../common/I18n';
+import getKebabCase from '../../../../../../common/utils/getKebabCase';
 import {useCustomerPortal} from '../../../../context';
-import getKebabCase from '../../../../utils/getKebabCase';
 
 const ActivationOutlet = () => {
 	const [{subscriptionGroups}] = useCustomerPortal();
